написал сервисы, которые будут работать с БД при помощи репозиториев а также использовать мапперы для преобразования объектов из DTO в сущности и обратно. Так же создал нужные таблицы

+++ src/main/java/ru/skypro/homework/repository/PhotoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.PhotoEntity;
 
 @Repository
-public interface PhotoRepository extends JpaRepository<PhotoEntity, Integer> {
+public interface PhotoRepository extends JpaRepository<PhotoEntity, Long> {
 }
